requirements.txt
pandas
numpy
flask
seaborn
scikit-learn==1.2.2
gunicorn
cassandra-driver

-e .
